uranai/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.4.0
Flask==3.0.3
frozenlist==1.4.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
numpy==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.11.1
